requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
folium>=0.15.0
streamlit-folium>=0.17.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
geopy>=2.4.0
